+++ kernel/kernel-api/src/main/java/io/delta/kernel/internal/SnapshotImpl.java
@@ -20,7 +20,6 @@ import static io.delta.kernel.internal.TableConfig.TOMBSTONE_RETENTION;
 
 import io.delta.kernel.ScanBuilder;
 import io.delta.kernel.Snapshot;
-import io.delta.kernel.engine.CommitCoordinatorClientHandler;
 import io.delta.kernel.engine.Engine;
 import io.delta.kernel.internal.actions.CommitInfo;
 import io.delta.kernel.internal.actions.DomainMetadata;
@@ -180,18 +179,6 @@ public class SnapshotImpl implements Snapshot {
    */
   public Optional<TableCommitCoordinatorClientHandler> getTableCommitCoordinatorClientHandlerOpt(
       Engine engine) {
-    return COORDINATED_COMMITS_COORDINATOR_NAME
-        .fromMetadata(metadata)
-        .map(
-            commitCoordinatorStr -> {
-              CommitCoordinatorClientHandler handler =
-                  engine.getCommitCoordinatorClientHandler(
-                      commitCoordinatorStr,
-                      COORDINATED_COMMITS_COORDINATOR_CONF.fromMetadata(metadata));
-              return new TableCommitCoordinatorClientHandler(
-                  handler,
-                  logPath.toString(),
-                  COORDINATED_COMMITS_TABLE_CONF.fromMetadata(metadata));
-            });
+    return Optional.empty();
   }
 }
