+++ kernel/kernel-api/src/main/java/io/delta/kernel/internal/coordinatedcommits/Commit.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.delta.kernel.engine.coordinatedcommits;
+package io.delta.kernel.internal.coordinatedcommits;
 
 import io.delta.kernel.annotation.Evolving;
 import io.delta.kernel.utils.FileStatus;
@@ -22,8 +22,8 @@ import io.delta.kernel.utils.FileStatus;
 /**
  * Representation of a commit file. It contains the version of the commit, the file status of the
  * commit, and the timestamp of the commit. This is used when we want to get the commit information
- * from the {@link io.delta.kernel.engine.CommitCoordinatorClientHandler#commit} and {@link
- * io.delta.kernel.engine.CommitCoordinatorClientHandler#getCommits} APIs.
+ * from the {@link CommitCoordinatorClientHandler#commit} and {@link
+ * CommitCoordinatorClientHandler#getCommits} APIs.
  *
  * @since 3.3.0
  */
