+++ kernel/kernel-defaults/src/main/java/io/delta/kernel/defaults/engine/DefaultEngine.java
@@ -16,7 +16,6 @@
 package io.delta.kernel.defaults.engine;
 
 import io.delta.kernel.engine.*;
-import java.util.Map;
 import org.apache.hadoop.conf.Configuration;
 
 /** Default implementation of {@link Engine} based on Hadoop APIs. */
@@ -45,12 +44,6 @@ public class DefaultEngine implements Engine {
   @Override
   public ParquetHandler getParquetHandler() {
     return new DefaultParquetHandler(hadoopConf);
-  }
-
-  @Override
-  public CommitCoordinatorClientHandler getCommitCoordinatorClientHandler(
-      String name, Map<String, String> conf) {
-    return new DefaultCommitCoordinatorClientHandler(hadoopConf, name, conf);
   }
 
   /**
