+++ kernel/kernel-defaults/src/test/scala/io/delta/kernel/defaults/LogReplayEngineMetricsSuite.scala
@@ -18,11 +18,12 @@ package io.delta.kernel.defaults
 
 import java.io.File
 import io.delta.kernel.Table
-import io.delta.kernel.engine.{CommitCoordinatorClientHandler, Engine, ExpressionHandler, FileSystemClient}
+import io.delta.kernel.engine.{Engine, ExpressionHandler, FileSystemClient}
 import io.delta.kernel.data.ColumnarBatch
 import io.delta.kernel.defaults.engine.{DefaultEngine, DefaultJsonHandler, DefaultParquetHandler}
 import io.delta.kernel.expressions.Predicate
 import io.delta.kernel.internal.checkpoints.Checkpointer
+import io.delta.kernel.internal.coordinatedcommits.CommitCoordinatorClientHandler
 import io.delta.kernel.internal.fs.Path
 import io.delta.kernel.internal.util.FileNames
 import io.delta.kernel.internal.util.Utils.toCloseableIterator
@@ -366,9 +367,6 @@ class MetricsEngine(config: Configuration) extends Engine {
   override def getFileSystemClient: FileSystemClient = impl.getFileSystemClient
 
   override def getParquetHandler: MetricsParquetHandler = parquetHandler
-
-  override def getCommitCoordinatorClientHandler(name: String, conf: util.Map[String, String]):
-  CommitCoordinatorClientHandler = impl.getCommitCoordinatorClientHandler(name, conf)
 }
 
 /**
